+++ jaorm-core/src/main/java/io/github/ulisse1996/jaorm/mapping/JaormCursor.java
@@ -3,7 +3,6 @@ package io.github.ulisse1996.jaorm.mapping;
 import io.github.ulisse1996.jaorm.ResultSetExecutor;
 import io.github.ulisse1996.jaorm.SqlExecutor;
 import io.github.ulisse1996.jaorm.SqlUtil;
-import io.github.ulisse1996.jaorm.entity.EntityDelegate;
 import io.github.ulisse1996.jaorm.exception.JaormSqlException;
 
 import java.sql.Connection;
@@ -12,23 +11,22 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
-import java.util.function.Supplier;
 
 public class JaormCursor<T> implements Cursor<T> {
 
-    private final Supplier<EntityDelegate<T>> supplier;
+    private final ThrowingFunction<ResultSet, T, SQLException> mapper;
     private final boolean fetched;
     private final Connection connection;
     private final PreparedStatement preparedStatement;
     private final ResultSetExecutor executor;
 
     public JaormCursor(Connection connection, PreparedStatement preparedStatement,
-                       SqlExecutor executor, Supplier<EntityDelegate<T>> supplier) {
+                       SqlExecutor executor, ThrowingFunction<ResultSet, T, SQLException> mapper) {
         this.fetched = false;
         this.connection = connection;
         this.preparedStatement = preparedStatement;
         this.executor = (ResultSetExecutor) executor;
-        this.supplier = supplier;
+        this.mapper = mapper;
     }
 
     @Override
@@ -46,26 +44,26 @@ public class JaormCursor<T> implements Cursor<T> {
         if (!this.fetched) {
             return new CursorIterator<>(
                     this.executor.getResultSet(),
-                    this.supplier,
+                    this.mapper,
                     this::close
             );
         } else {
-            throw new IllegalStateException("Cursor is already open !");
+            throw new IllegalStateException("Cursor is already consumed !");
         }
     }
 
     private static class CursorIterator<T> implements Iterator<T> {
 
         private final ResultSet resultSet;
-        private final Supplier<EntityDelegate<T>> supplier;
+        private final ThrowingFunction<ResultSet, T, SQLException> mapper;
         private final Runnable onClose;
         private boolean hasNext;
 
-        CursorIterator(ResultSet rs, Supplier<EntityDelegate<T>> supplier,
+        CursorIterator(ResultSet rs, ThrowingFunction<ResultSet, T, SQLException> mapper,
                        Runnable onClose) {
             this.resultSet = rs;
             this.hasNext = canGoNext(rs);
-            this.supplier = supplier;
+            this.mapper = mapper;
             this.onClose = onClose;
         }
 
@@ -85,12 +83,9 @@ public class JaormCursor<T> implements Cursor<T> {
             return next;
         }
 
-        @SuppressWarnings("unchecked")
         private T mapNext() {
             try {
-                EntityDelegate<T> entity = supplier.get();
-                entity.setEntity(resultSet);
-                return (T) entity;
+                return mapper.apply(this.resultSet);
             } catch (SQLException ex) {
                 this.onClose.run();
                 throw new JaormSqlException(ex);